backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Authlib==1.4.1
cachetools==5.5.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
cryptography==43.0.1
distlib==0.3.8
dj-rest-auth==6.0.0
Django==5.1.6
django-allauth==65.4.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
dparse==0.6.4
drf-yasg==1.21.7
Faker==33.1.0
filelock==3.16.1
google-auth==2.35.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
Jinja2==3.1.5
jwt==1.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
packaging==24.1
pbr==6.0.0
platformdirs==4.1.0
psutil==6.1.1
psycopg==3.2.2
psycopg-binary==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safety==3.2.14
safety-schemas==0.0.10
shellingham==1.5.4
six==1.16.0
sqlparse==0.5.1
stevedore==5.1.0
tomli==2.2.1
typer==0.15.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.29.2
virtualenv-clone==0.5.7
virtualenvwrapper==6.0.0
